improve the `remove` unittest
It was pretty terrible before, and let a bug slip through.
